+++ fdbclient/ClientKnobs.cpp
@@ -206,10 +206,6 @@ void ClientKnobs::initialize(Randomize randomize) {
 
 	init( IS_ACCEPTABLE_DELAY,                     1.5 );
 
-	init( HTTP_READ_SIZE,                     128*1024 );
-	init( HTTP_SEND_SIZE,                      32*1024 );
-	init( HTTP_VERBOSE_LEVEL,                        0 );
-	init( HTTP_REQUEST_ID_HEADER,                   "" );
 	init( HTTP_REQUEST_AWS_V4_HEADER,             true );
 	init( BLOBSTORE_ENCRYPTION_TYPE,                "" );
 	init( BLOBSTORE_CONNECT_TRIES,                  10 );
