+++ fdbrpc/HTTP.actor.cpp
@@ -18,9 +18,11 @@
  * limitations under the License.
  */
 
-#include "fdbclient/HTTP.h"
-#include "fdbclient/md5/md5.h"
-#include "fdbclient/libb64/encode.h"
+#include "fdbrpc/HTTP.h"
+
+#include "fdbrpc/md5/md5.h"
+#include "fdbrpc/libb64/encode.h"
+#include "flow/Knobs.h"
 #include <cctype>
 #include "flow/actorcompiler.h" // has to be last include
 
@@ -149,7 +151,7 @@ ACTOR Future<size_t> read_delimited_into_string(Reference<IConnection> conn,
 		// Next search will start at the current end of the buffer - delim size + 1
 		if (sPos >= lookBack)
 			sPos -= lookBack;
-		wait(success(read_into_string(conn, buf, CLIENT_KNOBS->HTTP_READ_SIZE)));
+		wait(success(read_into_string(conn, buf, FLOW_KNOBS->HTTP_READ_SIZE)));
 	}
 }
 
@@ -157,7 +159,7 @@ ACTOR Future<size_t> read_delimited_into_string(Reference<IConnection> conn,
 ACTOR Future<Void> read_fixed_into_string(Reference<IConnection> conn, int len, std::string* buf, size_t pos) {
 	state int stop_size = pos + len;
 	while (buf->size() < stop_size)
-		wait(success(read_into_string(conn, buf, CLIENT_KNOBS->HTTP_READ_SIZE)));
+		wait(success(read_into_string(conn, buf, FLOW_KNOBS->HTTP_READ_SIZE)));
 	return Void();
 }
 
@@ -348,7 +350,7 @@ ACTOR Future<Reference<HTTP::Response>> doRequest(Reference<IConnection> conn,
 	// There is no standard http request id header field, so either a global default can be set via a knob
 	// or it can be set per-request with the requestIDHeader argument (which overrides the default)
 	if (requestIDHeader.empty()) {
-		requestIDHeader = CLIENT_KNOBS->HTTP_REQUEST_ID_HEADER;
+		requestIDHeader = FLOW_KNOBS->HTTP_REQUEST_ID_HEADER;
 	}
 
 	state bool earlyResponse = false;
@@ -380,13 +382,13 @@ ACTOR Future<Reference<HTTP::Response>> doRequest(Reference<IConnection> conn,
 		// Prepend headers to content packer buffer chain
 		pContent->prependWriteBuffer(pFirst, pLast);
 
-		if (CLIENT_KNOBS->HTTP_VERBOSE_LEVEL > 1)
+		if (FLOW_KNOBS->HTTP_VERBOSE_LEVEL > 1)
 			printf("[%s] HTTP starting %s %s ContentLen:%d\n",
 			       conn->getDebugID().toString().c_str(),
 			       verb.c_str(),
 			       resource.c_str(),
 			       contentLen);
-		if (CLIENT_KNOBS->HTTP_VERBOSE_LEVEL > 2) {
+		if (FLOW_KNOBS->HTTP_VERBOSE_LEVEL > 2) {
 			for (auto h : headers)
 				printf("Request Header: %s: %s\n", h.first.c_str(), h.second.c_str());
 		}
@@ -407,7 +409,7 @@ ACTOR Future<Reference<HTTP::Response>> doRequest(Reference<IConnection> conn,
 				break;
 			}
 
-			state int trySend = CLIENT_KNOBS->HTTP_SEND_SIZE;
+			state int trySend = FLOW_KNOBS->HTTP_SEND_SIZE;
 			wait(sendRate->getAllowance(trySend));
 			int len = conn->write(pContent->getUnsent(), trySend);
 			if (pSent != nullptr)
@@ -461,7 +463,7 @@ ACTOR Future<Reference<HTTP::Response>> doRequest(Reference<IConnection> conn,
 			}
 		}
 
-		if (CLIENT_KNOBS->HTTP_VERBOSE_LEVEL > 0) {
+		if (FLOW_KNOBS->HTTP_VERBOSE_LEVEL > 0) {
 			printf("[%s] HTTP %scode=%d early=%d, time=%fs %s %s contentLen=%d [%d out, response content len %d]\n",
 			       conn->getDebugID().toString().c_str(),
 			       (err.present() ? format("*ERROR*=%s ", err.get().name()).c_str() : ""),
@@ -474,7 +476,7 @@ ACTOR Future<Reference<HTTP::Response>> doRequest(Reference<IConnection> conn,
 			       total_sent,
 			       (int)r->contentLen);
 		}
-		if (CLIENT_KNOBS->HTTP_VERBOSE_LEVEL > 2) {
+		if (FLOW_KNOBS->HTTP_VERBOSE_LEVEL > 2) {
 			printf("[%s] HTTP RESPONSE:  %s %s\n%s\n",
 			       conn->getDebugID().toString().c_str(),
 			       verb.c_str(),
@@ -490,7 +492,7 @@ ACTOR Future<Reference<HTTP::Response>> doRequest(Reference<IConnection> conn,
 	} catch (Error& e) {
 		double elapsed = timer() - send_start;
 		// A bad_request_id error would have already been logged in verbose mode before err is thrown above.
-		if (CLIENT_KNOBS->HTTP_VERBOSE_LEVEL > 0 && e.code() != error_code_http_bad_request_id) {
+		if (FLOW_KNOBS->HTTP_VERBOSE_LEVEL > 0 && e.code() != error_code_http_bad_request_id) {
 			printf("[%s] HTTP *ERROR*=%s early=%d, time=%fs %s %s contentLen=%d [%d out]\n",
 			       conn->getDebugID().toString().c_str(),
 			       e.name(),
