+++ fdbrpc/libb64/cdecode.c
