+++ fdbrpc/libb64/cencode.c
