+++ fdbrpc/libb64/cencode.h
