+++ fdbrpc/libb64/decode.h
