+++ fdbrpc/libb64/encode.h
