+++ flow/Knobs.h
@@ -233,6 +233,13 @@ public:
 	int LOW_PRIORITY_DELAY_COUNT;
 	double LOW_PRIORITY_MAX_DELAY;
 
+	// HTTP
+	int HTTP_READ_SIZE;
+	int HTTP_SEND_SIZE;
+	int HTTP_VERBOSE_LEVEL;
+	std::string HTTP_REQUEST_ID_HEADER;
+	bool HTTP_RESPONSE_SKIP_VERIFY_CHECKSUM_FOR_PARTIAL_CONTENT; // skip verify md5 checksum for 206 response
+
 	// IAsyncFile
 	int64_t INCREMENTAL_DELETE_TRUNCATE_AMOUNT;
 	double INCREMENTAL_DELETE_INTERVAL;
